+++ web/default/src/components/data-table/index.ts
@@ -18,7 +18,6 @@ For commercial licensing, please contact support@quantumnous.com
 */
 export { DataTablePagination } from './pagination'
 export { DataTableColumnHeader } from './column-header'
-export { DataTableFacetedFilter } from './faceted-filter'
 export { DataTableViewOptions } from './view-options'
 export { DataTableToolbar } from './toolbar'
 export { DataTableBulkActions } from './bulk-actions'
@@ -34,13 +33,10 @@ export {
 } from '@/components/ui/table'
 export {
   StaticDataTable,
-  StaticDataTableEmptyState,
   StaticDataTableEmptyRow,
   staticDataTableClassNames,
-  type StaticDataTableElement,
 } from './static-data-table'
 export {
-  DataTableHeader,
   DataTableRow,
   DataTableView,
   type DataTableColumnClassName,
@@ -48,15 +44,9 @@ export {
   type DataTableRenderRowHelpers,
   type DataTableViewProps,
 } from './data-table-view'
-export { TableSkeleton } from './table-skeleton'
-export { TableEmpty } from './table-empty'
 export { MobileCardList } from './mobile-card-list'
 export { DataTablePage, type DataTablePageProps } from './data-table-page'
-export {
-  useDataTable,
-  type DataTableRowSelectionPredicate,
-  type UseDataTableOptions,
-} from './use-data-table'
+export { useDataTable } from './use-data-table'
 
 export const DISABLED_ROW_DESKTOP =
   'bg-muted/85 hover:bg-muted [&>td:first-child]:border-l-muted-foreground/35 [&>td:first-child]:border-l-4 [&>td:first-child]:pl-1'
