+++ web/default/src/features/system-settings/models/upstream-ratio-sync-table.tsx
@@ -29,7 +29,7 @@ import {
   SelectValue,
 } from '@/components/ui/select'
 import { DataTableView, useDataTable } from '@/components/data-table'
-import { DataTablePagination } from '@/components/data-table/pagination'
+import { DataTablePagination } from '@/components/data-table'
 import type { DifferencesMap, RatioType } from '../types'
 import { RATIO_TYPE_OPTIONS } from './constants'
 import { useUpstreamRatioSyncColumns } from './upstream-ratio-sync-columns'
